refactor: split up big storage service in smaller sub services
